rename typography folder
